3.0 Extract storage management and background writer systems
